+++ events/threadDelete.js
@@ -1,0 +1,9 @@
+const { removeThread } = require("../utils/threadActions")
+
+const run = ( client ) => {
+    client.on("threadDelete", (thread) => {
+        removeThread(thread.id)
+    })
+}
+
+module.exports = { run }
